refactor: 접근제한자 변경 및 패키지 분리
